Add task 1ans

+++ 0x0B-malloc_free/1-strdup.c
@@ -9,7 +9,7 @@
  */
 char *_strdup(char *str)
 {
-	int i;
+	long unsigned int i;
 	
 	char *arr = malloc(sizeof(int) * strlen(str));
